+++ NEWS.md
@@ -1,3 +1,24 @@
+# bbr 1.1.0
+
+## New features and changes
+
+* Added `tags_diff()` function for comparing the tags between different models. (#337)
+
+* Added `model_diff()` function for comparing the model files between different models. (#342)
+
+* Added `check_up_to_date()` function for checking whether the model file(s) and data file(s) associated with a model have changed since the model was run.  (#338)
+
+* Added more documentation about the heuristics returned from `model_summary.bbi_nonmem_model()` (#343)
+
+## Bug fixes
+
+* `param_estimates()` now correctly errors when a Bayesian method is used but is _not_ the final method. (#344)
+
+## Developer-facing changes
+
+* Added a `bbi_model` parent class to `bbi_nonmem_model` and `bbi_nonmem_summary` objects. Many of the helpers in `get-path-from-object.R` now dispatch on this class. This had been discussed in the past but was primarily done now in preparation for beginning development for Stan modeling, which will create `bbi_stan_model` and `bbi_stan_summary` objects that will also inherit from this parent class. (#332)
+
+
 # bbr 1.0.0
 
 This release is fairly small in terms of changes, but it increments to a new major release version primarily because of the name change which happened in `bbr 0.12.0`. The most significant change, from a user perspective, is to the default behavior of where `bbr` looks for `bbi` on the system. This change is described in [issue #321](https://github.com/metrumresearchgroup/bbr/issues/321) and a bit more detail is given below.